Added functionality to check that the gem has been published to the specs.*.gz file on the git hub gems server.
This file is updated every 20 mins (on the hour, 20 past and 20 to the hour), so will let you know when it's there.

The rubygems caches the contents of the file, and uses if-modified-since to make sure we don't swamp the server.

That said, we could be a little more cache aware if it causes issues.